requirements.txt
Flask
flask-login
flask-cors
PyJWT
flask-limiter
stripe
gunicorn